Refactor integrations setup for easier separation of different setup options (#1738)
* Refactor addIntegrationRequest params to object


* Move SetupIntegrationFormInputs to own file


* Split form inputs into more sections visually


* Split form inputs into more sections logically


* Minor copy update for checkpoint location


* Update toggleworkflow method per Ryan's feedback


---------
